requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
celery==5.4.0
redis==5.0.7
django-redis==5.4.0
psycopg2-binary==2.9.9
django-filter==24.2
python-decouple==3.8
django-silk==5.1.0

# TESTS
pytest==8.2.2
pytest-django==4.8.0
pytest-mock==3.14.0
fakeredis==2.23.2